+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## Retail: [4.82] ##
+## Retail: [4.83] ##
 ### Changes ###
 
-  * Fix a lot of green icons
-  * Real fix for StaticPopups
+  * Fix Clique icon texture
+  * Fix Reminder for real. Thanks @kodewdle (Simpy)
+  * Fix Merchant Frame option
+  * Change the Option to toggle the Guild on the MicroBar
   
 
 ## Classic: [1.54] ##